split gui1 font interface into a font_config and sdl_ttf interface
move all of these into font folder
