+++ src/editor/palette/tristate_button.cpp
@@ -18,7 +18,7 @@
 
 #include "tristate_button.hpp"
 
-#include "font.hpp"
+#include "font/sdl_ttf.hpp"
 #include "game_config.hpp"
 #include "image.hpp"
 #include "log.hpp"
