+++ src/font/text_cache.cpp
@@ -1,0 +1,41 @@
+
+#include "text_cache.hpp"
+
+namespace font {
+
+text_cache::text_list text_cache::cache_;
+
+unsigned int text_cache::max_size_ = 50;
+
+void text_cache::resize(unsigned int size)
+{
+	// DBG_FT << "Text cache: resize from: " << max_size_ << " to: "
+	// 	<< size << " items in cache: " << cache_.size() << '\n';
+
+	while(size < cache_.size()) {
+		cache_.pop_back();
+	}
+	max_size_ = size;
+}
+
+text_surface &text_cache::find(text_surface const &t)
+{
+	static size_t lookup_ = 0, hit_ = 0;
+	text_list::iterator it_bgn = cache_.begin(), it_end = cache_.end();
+	text_list::iterator it = std::find(it_bgn, it_end, t);
+	if (it != it_end) {
+		cache_.splice(it_bgn, cache_, it);
+		++hit_;
+	} else {
+		if (cache_.size() >= max_size_)
+			cache_.pop_back();
+		cache_.push_front(t);
+	}
+	if (++lookup_ % 1000 == 0) {
+		// DBG_FT << "Text cache: " << lookup_ << " lookups, " << (hit_ / 10) << "% hits\n";
+		hit_ = 0;
+	}
+	return cache_.front();
+}
+
+} // end namespace font
