+++ src/font/text_surface.hpp
@@ -1,0 +1,88 @@
+/*
+   Copyright (C) 2016 by Chris Beck<render787@gmail.com>
+   Part of the Battle for Wesnoth Project http://www.wesnoth.org/
+
+   This program is free software; you can redistribute it and/or modify
+   it under the terms of the GNU General Public License as published by
+   the Free Software Foundation; either version 2 of the License, or
+   (at your option) any later version.
+   This program is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY.
+
+   See the COPYING file for more details.
+*/
+
+#ifndef FONT_TEXT_SURFACE_HPP
+#define FONT_TEXT_SURFACE_HPP
+
+#include "font_id.hpp" // for text_chunk
+#include "sdl/utils.hpp"
+#include "sdl/image.hpp"
+
+#include <SDL_ttf.h>
+
+#include <string>
+#include <vector>
+
+/***
+ * Note: This is specific to the SDL_TTF codepath.
+ */
+
+namespace font {
+
+const SDL_Color NORMAL_COLOR = {0xDD,0xDD,0xDD,0},
+                GRAY_COLOR   = {0x77,0x77,0x77,0},
+                LOBBY_COLOR  = {0xBB,0xBB,0xBB,0},
+                GOOD_COLOR   = {0x00,0xFF,0x00,0},
+                BAD_COLOR    = {0xFF,0x00,0x00,0},
+                BLACK_COLOR  = {0x00,0x00,0x00,0},
+                YELLOW_COLOR = {0xFF,0xFF,0x00,0},
+                BUTTON_COLOR = {0xBC,0xB0,0x88,0},
+                PETRIFIED_COLOR = {0xA0,0xA0,0xA0,0},
+                TITLE_COLOR  = {0xBC,0xB0,0x88,0},
+				LABEL_COLOR  = {0x6B,0x8C,0xFF,0},
+				BIGMAP_COLOR = {0xFF,0xFF,0xFF,0};
+const SDL_Color DISABLED_COLOR = inverse(PETRIFIED_COLOR);
+
+static const size_t max_text_line_width = 4096;
+
+class text_surface
+{
+public:
+	text_surface(std::string const &str, int size, SDL_Color color, int style);
+	text_surface(int size, SDL_Color color, int style);
+	void set_text(std::string const &str);
+
+	void measure() const;
+	size_t width() const;
+	size_t height() const;
+#ifdef	HAVE_FRIBIDI
+	bool is_rtl() const { return is_rtl_; }	// Right-To-Left alignment
+#endif
+	std::vector<surface> const & get_surfaces() const;
+
+	bool operator==(text_surface const &t) const {
+		return hash_ == t.hash_ && font_size_ == t.font_size_
+			&& color_ == t.color_ && style_ == t.style_ && str_ == t.str_;
+	}
+	bool operator!=(text_surface const &t) const { return !operator==(t); }
+private:
+	int hash_;
+	int font_size_;
+	SDL_Color color_;
+	int style_;
+	mutable int w_, h_;
+	std::string str_;
+	mutable bool initialized_;
+	mutable std::vector<text_chunk> chunks_;
+	mutable std::vector<surface> surfs_;
+#ifdef	HAVE_FRIBIDI
+	bool is_rtl_;
+	void bidi_cvt();
+#endif
+	void hash();
+};
+
+} // end namespace font
+
+#endif
