+++ src/game_launcher.hpp
@@ -18,7 +18,7 @@
 
 #include "editor/editor_main.hpp"       // for EXIT_STATUS
 #include "events.hpp"                   // for event_context
-#include "font.hpp"                     // for manager
+#include "font/font_config.hpp"         // for manager
 #include "game_preferences.hpp"         // for manager
 #include "hotkey/hotkey_manager.hpp"    // for manager
 #include "image.hpp"                    // for manager
