+++ src/reports.cpp
@@ -18,7 +18,7 @@
 #include "attack_prediction.hpp"
 //#include "editor/editor_controller.hpp"
 //#include "editor/palette/terrain_palettes.hpp"
-#include "font.hpp"
+#include "font/sdl_ttf.hpp"
 #include "game_preferences.hpp"
 #include "gettext.hpp"
 #include "language.hpp"
