+++ src/wesnoth.cpp
@@ -22,7 +22,7 @@
 #include "editor/editor_main.hpp"
 #include "filesystem.hpp"               // for filesystem::file_exists, filesystem::io_exception, etc
 #include "floating_label.hpp"
-#include "font.hpp"                     // for load_font_config, etc
+#include "font/font_config.hpp"         // for load_font_config, etc
 #include "formula/formula.hpp"                  // for formula_error
 #include "game_config.hpp"              // for path, debug, debug_lua, etc
 #include "game_config_manager.hpp"      // for game_config_manager, etc
