+++ src/widgets/slider.cpp
@@ -18,7 +18,7 @@
 
 #include "widgets/slider.hpp"
 #include "game_config.hpp"
-#include "font.hpp"
+#include "font/sdl_ttf.hpp"
 #include "image.hpp"
 #include "sdl/rect.hpp"
 #include "sound.hpp"
